move add_observation_table from base mixin into Context api object
